Revert "fix build errors"
This reverts commit f4acd6e9776c899a0523df10b64e55783b44c451.

+++ beacon-chain/core/epoch/precompute/reward_penalty.go
@@ -22,7 +22,7 @@ func ProcessRewardsAndPenaltiesPrecompute(
 
 	numOfVals := state.NumValidators()
 	// Guard against an out-of-bounds using validator balance precompute.
-	if uint64(len(vp)) != numOfVals || len(vp) != state.BalancesLength() {
+	if len(vp) != numOfVals || len(vp) != state.BalancesLength() {
 		return state, errors.New("precomputed registries not the same length as state registries")
 	}
 
@@ -35,7 +35,7 @@ func ProcessRewardsAndPenaltiesPrecompute(
 		return nil, errors.Wrap(err, "could not get attestation delta")
 	}
 	validatorBals := state.Balances()
-	for i := uint64(0); i < numOfVals; i++ {
+	for i := 0; i < numOfVals; i++ {
 		vp[i].BeforeEpochTransitionBalance = validatorBals[i]
 
 		// Compute the post balance of the validator after accounting for the
